+++ src/libgeometry/shape_info.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <libgeometry/geometry.h>
+#include <src/libgeometry/geometry.h>
 
-float get_perimeter(circle*);
-float get_area(circle*);
+double get_perimeter(circle*);
+double get_area(circle*);
